feat(92.1-01): move search transport DTOs into uc-daemon-contract
- Create uc-daemon-contract/src/api/dto/search.rs with SearchQueryResponse, SearchStatusResponse, SearchRebuildAcceptedResponse structs
- Add pub mod search to uc-daemon-contract/src/api/dto/mod.rs
- Replace daemon-local search.rs with thin re-export (pub use uc_daemon_contract::api::dto::search::*)
- cargo check -p uc-daemon-contract -p uc-daemon passes
