+++ src-tauri/crates/uc-daemon-contract/src/api/dto/mod.rs
@@ -2,6 +2,7 @@ pub mod clipboard;
 pub mod device;
 pub mod encryption;
 pub mod pairing;
+pub mod search;
 pub mod settings;
 pub mod setup;
 pub mod ws;
