+++ src-tauri/crates/uc-daemon/src/api/dto/search.rs
@@ -1,64 +1,6 @@
-//! Transport DTOs for the search query, status, and rebuild HTTP endpoints (Phase 92).
-
-use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
-use uc_core::search::FileType;
-
-/// Single search result — mirrors `SearchResult` with camelCase transport names.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchResultDto {
-    pub entry_id: String,
-    pub file_type: FileType,
-    pub active_time_ms: i64,
-    pub text_preview: Option<String>,
-    pub mime_type: String,
-    pub file_extensions: Vec<String>,
-}
-
-/// Response envelope for `GET /search/query`.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchQueryResponse {
-    pub data: Vec<SearchResultDto>,
-    pub total: u32,
-    pub has_more: bool,
-    pub ts: i64,
-}
-
-/// Search index availability snapshot embedded in `SearchStatusResponse`.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchStatusData {
-    /// One of: "ready", "rebuilding", "unavailable"
-    pub state: String,
-    /// Optional reason code (present when rebuilding or unavailable).
-    pub reason: Option<String>,
-    /// Millisecond timestamp of the last rebuild start (from `SearchIndexMeta`).
-    pub last_rebuild_started_at_ms: Option<i64>,
-    /// Millisecond timestamp of the last completed rebuild (from `SearchIndexMeta`).
-    pub last_rebuild_completed_at_ms: Option<i64>,
-}
-
-/// Response envelope for `GET /search/status`.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchStatusResponse {
-    pub data: SearchStatusData,
-    pub ts: i64,
-}
-
-/// Data payload inside `SearchRebuildAcceptedResponse`.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchRebuildAcceptedData {
-    pub accepted: bool,
-}
-
-/// Response envelope for `POST /search/rebuild`.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchRebuildAcceptedResponse {
-    pub data: SearchRebuildAcceptedData,
-    pub ts: i64,
-}
+//! Search transport DTOs — re-exported from the shared daemon-contract crate.
+//!
+//! The single source of truth for search response envelopes lives in
+//! `uc-daemon-contract`. This module is a thin pass-through so all existing
+//! daemon code at `crate::api::dto::search::*` continues to resolve.
+pub use uc_daemon_contract::api::dto::search::*;
